Refactor: Simplify project structure and remove dead code
- Deleted `src/experimental` module and associated CLI `highlight` command (unused feature).
- Flattened `src/ast/` into `src/ast.rs`.
- Flattened `src/grammar/` into `src/grammar.rs` (moved `glossa.pest` to `src/grammar.pest`).
- Flattened `src/errors/` into `src/errors.rs`.
- Fixed `is_std_type` in `codegen/rust.rs` to include `GlossaType::Unknown`, preventing incorrect `g_` prefixing on standard iterator methods.
