--- src/experimental/bard.rs
@@ -1,809 +1,0 @@
-//! The Bard (ὁ Ῥαψῳδός) - Semantic Syntax Highlighter
-//!
-//! This module implements a semantic syntax highlighter that colors the source code
-//! based on the grammatical role of each word (Subject, Object, Verb, etc.).
-//!
-//! # Philosophy
-//!
-//! Unlike traditional syntax highlighters that use regexes, The Bard uses the
-//! compiler's own morphological analysis to understand the code.
-//!
-//! * **Nominative (Subject)**: Blue (The agent/foundation)
-//! * **Accusative (Object)**: Red (The target of action)
-//! * **Dative (Indirect)**: Yellow (The recipient)
-//! * **Genitive (Possession)**: Magenta (Ownership)
-//! * **Verb (Action)**: Green (Go!)
-//! * **Adjective**: Cyan (Modification)
-//! * **Literals**: Italic/White
-//!
-//! # Usage
-//!
-//! ```rust
-//! use glossa::experimental::bard::highlight;
-//!
-//! let source = "ὁ ἄνθρωπος τὸν λόγον λέγει.";
-//! let highlighted = highlight(source).unwrap();
-//! println!("{}", highlighted);
-//! ```
-
-use crossterm::style::Stylize;
-use std::fmt::Write;
-
-use crate::ast::{
-    BinOperator, Clause, Expr, Program, Statement, TestDecl, TraitDef, TraitImplDef, TypeDef,
-    UnaryOperator, Word,
-};
-use crate::errors::GlossaError;
-use crate::morphology::{
-    Case, DisambiguationContext, PartOfSpeech, analyze_article, analyze_participle, resolve_best,
-};
-use crate::parser::parse;
-
-/// Highlight the source code with semantic colors
-pub fn highlight(source: &str) -> Result<String, GlossaError> {
-    let program = parse(source)?;
-    let mut highlighter = Highlighter::new();
-    highlighter.highlight_program(&program)?;
-    Ok(highlighter.output)
-}
-
-struct Highlighter {
-    output: String,
-    context: DisambiguationContext,
-}
-
-impl Highlighter {
-    fn new() -> Self {
-        Self {
-            output: String::new(),
-            context: DisambiguationContext::new(),
-        }
-    }
-
-    fn highlight_program(&mut self, program: &Program) -> Result<(), GlossaError> {
-        for (i, stmt) in program.statements.iter().enumerate() {
-            if i > 0 {
-                self.output.push('\n');
-            }
-            self.highlight_statement(stmt)?;
-        }
-        Ok(())
-    }
-
-    fn highlight_statement(&mut self, stmt: &Statement) -> Result<(), GlossaError> {
-        match stmt {
-            Statement::Regular {
-                clauses,
-                is_query,
-                is_propagate,
-            } => {
-                for (i, clause) in clauses.iter().enumerate() {
-                    if i > 0 {
-                        self.output.push_str(", ");
-                    }
-                    self.highlight_clause(clause)?;
-                }
-
-                if *is_query {
-                    self.output.push('?');
-                } else if *is_propagate {
-                    self.output.push(';');
-                } else {
-                    self.output.push('.');
-                }
-            }
-            Statement::TypeDefinition(def) => self.highlight_type_def(def)?,
-            Statement::TraitDefinition(def) => self.highlight_trait_def(def)?,
-            Statement::TraitImpl(def) => self.highlight_trait_impl(def)?,
-            Statement::TestDeclaration(decl) => self.highlight_test_decl(decl)?,
-        }
-        Ok(())
-    }
-
-    fn highlight_clause(&mut self, clause: &Clause) -> Result<(), GlossaError> {
-        for (i, expr) in clause.expressions.iter().enumerate() {
-            if i > 0 {
-                self.output.push(' ');
-            }
-            self.highlight_expr(expr)?;
-        }
-        Ok(())
-    }
-
-    fn highlight_expr(&mut self, expr: &Expr) -> Result<(), GlossaError> {
-        match expr {
-            Expr::StringLiteral(s) => {
-                write!(self.output, "«{}»", s.as_str().italic()).unwrap();
-            }
-            Expr::NumberLiteral(n) => {
-                write!(self.output, "{}", n.to_string().italic()).unwrap();
-            }
-            Expr::BooleanLiteral(b) => {
-                let s = if *b { "ἀληθές" } else { "ψεῦδος" };
-                write!(self.output, "{}", s.italic()).unwrap();
-            }
-            Expr::Word(w) => self.highlight_word(w)?,
-            Expr::Phrase(terms) => {
-                for (i, term) in terms.iter().enumerate() {
-                    if i > 0 {
-                        self.output.push(' ');
-                    }
-                    self.highlight_expr(term)?;
-                }
-            }
-            Expr::PropertyAccess { owner, property } => {
-                self.highlight_expr(owner)?;
-                self.output.push(' ');
-                self.highlight_expr(property)?;
-            }
-            Expr::Call { verb, arguments } => {
-                // Highlight verb
-                self.highlight_word(verb)?;
-                // Arguments
-                for arg in arguments {
-                    self.output.push(' ');
-                    self.highlight_expr(arg)?;
-                }
-            }
-            Expr::Binding { name, value } => {
-                self.highlight_word(name)?;
-                self.output.push(' ');
-                self.highlight_expr(value)?;
-                self.output.push(' ');
-                write!(self.output, "{}", "ἔστω".bold()).unwrap();
-            }
-            Expr::BinOp { left, op, right } => {
-                self.highlight_expr(left)?;
-                self.output.push(' ');
-                self.highlight_binop(op);
-                self.output.push(' ');
-                self.highlight_expr(right)?;
-            }
-            Expr::UnaryOp { op, operand } => {
-                match op {
-                    UnaryOperator::Unwrap => {
-                        self.highlight_expr(operand)?;
-                        write!(self.output, "{}", "!".bold().red()).unwrap();
-                    }
-                    UnaryOperator::Not => {
-                        write!(self.output, "{}", "οὐ".bold()).unwrap(); // Simplified
-                        self.output.push(' ');
-                        self.highlight_expr(operand)?;
-                    }
-                    UnaryOperator::Neg => {
-                        write!(self.output, "-").unwrap();
-                        self.highlight_expr(operand)?;
-                    }
-                }
-            }
-            Expr::Block(stmts) => {
-                self.output.push_str("{ ");
-                for (i, stmt) in stmts.iter().enumerate() {
-                    if i > 0 {
-                        self.output.push(' ');
-                    }
-                    self.highlight_statement(stmt)?;
-                }
-                self.output.push_str(" }");
-            }
-            Expr::ArrayLiteral(elements) => {
-                self.output.push('[');
-                for (i, el) in elements.iter().enumerate() {
-                    if i > 0 {
-                        self.output.push_str(", ");
-                    }
-                    self.highlight_expr(el)?;
-                }
-                self.output.push(']');
-            }
-            Expr::IndexAccess { array, index } => {
-                self.highlight_expr(array)?;
-                self.output.push('[');
-                self.highlight_expr(index)?;
-                self.output.push(']');
-            }
-        }
-        Ok(())
-    }
-
-    fn highlight_word(&mut self, w: &Word) -> Result<(), GlossaError> {
-        // 1. Check for article (sets context)
-        if let Some(ctx) = analyze_article(&w.original) {
-            self.context = ctx;
-            write!(self.output, "{}", w.original).unwrap(); // Articles plain or dim? Let's leave plain
-            return Ok(());
-        }
-
-        // 2. Check for participle
-        let in_lexicon = crate::morphology::lexicon::lookup(&w.normalized).is_some();
-        let is_numeral = crate::morphology::lexicon::numeral_value(&w.normalized).is_some();
-
-        if !in_lexicon && !is_numeral && analyze_participle(&w.normalized).is_some() {
-            write!(self.output, "{}", w.original.cyan()).unwrap(); // Participles as cyan (adjectival)
-            return Ok(());
-        }
-
-        // 3. Analyze and disambiguate
-        let analyses = crate::morphology::analyze_all(&w.normalized);
-        let best = resolve_best(analyses, &self.context);
-
-        // Update context if it's a verb
-        if best.part_of_speech == PartOfSpeech::Verb {
-            self.context = DisambiguationContext::from_verb(&best);
-        } else {
-            // Consume context for nouns
-            self.context = DisambiguationContext::new();
-        }
-
-        // 4. Apply Color
-        let styled = match best.part_of_speech {
-            PartOfSpeech::Verb => w.original.green().bold(),
-            PartOfSpeech::Noun | PartOfSpeech::Pronoun => match best.case {
-                Some(Case::Nominative) => w.original.blue().bold(),
-                Some(Case::Accusative) => w.original.red(),
-                Some(Case::Dative) => w.original.yellow(),
-                Some(Case::Genitive) => w.original.magenta(),
-                Some(Case::Vocative) => w.original.blue().italic(), // Vocative as blue italic
-                None => w.original.white(),
-            },
-            PartOfSpeech::Adjective => w.original.cyan(),
-            PartOfSpeech::Preposition => w.original.white().bold(),
-            PartOfSpeech::Conjunction => w.original.white().bold(),
-            PartOfSpeech::Numeral => w.original.italic(),
-            _ => w.original.white(), // Default
-        };
-
-        write!(self.output, "{}", styled).unwrap();
-        Ok(())
-    }
-
-    fn highlight_binop(&mut self, op: &BinOperator) {
-        let s = match op {
-            BinOperator::Add => "+",
-            BinOperator::Sub => "-",
-            BinOperator::Mul => "*",
-            BinOperator::Div => "/",
-            BinOperator::Mod => "%",
-            BinOperator::Eq => "==",
-            BinOperator::Ne => "!=",
-            BinOperator::Lt => "<",
-            BinOperator::Le => "<=",
-            BinOperator::Gt => ">",
-            BinOperator::Ge => ">=",
-            BinOperator::And => "&&",
-            BinOperator::Or => "||",
-        };
-        write!(self.output, "{}", s.bold()).unwrap();
-    }
-
-    // --- Definitions (Simplified highlighting for now) ---
-
-    fn highlight_type_def(&mut self, def: &TypeDef) -> Result<(), GlossaError> {
-        write!(
-            self.output,
-            "{} {} {} {{ ... }}",
-            "εἶδος".bold(),
-            def.name.original.blue().bold(),
-            "ὁρίζειν".bold()
-        )
-        .unwrap();
-        Ok(())
-    }
-
-    fn highlight_trait_def(&mut self, def: &TraitDef) -> Result<(), GlossaError> {
-        write!(
-            self.output,
-            "{} {} {} {{ ... }}",
-            "χαρακτήρ".bold(),
-            def.name.original.blue().bold(),
-            "ὁρίζειν".bold()
-        )
-        .unwrap();
-        Ok(())
-    }
-
-    fn highlight_trait_impl(&mut self, def: &TraitImplDef) -> Result<(), GlossaError> {
-        write!(
-            self.output,
-            "{} {} {} {} {{ ... }}",
-            "εἶδος".bold(),
-            def.type_name.original.blue().bold(),
-            "τῷ".white(),
-            def.trait_name.original.cyan(),
-            // Missing implementation keyword? Syntax is `εἶδος Type τῷ Trait ἐμπίπτειν`
-        )
-        .unwrap();
-        Ok(())
-    }
-
-    fn highlight_test_decl(&mut self, decl: &TestDecl) -> Result<(), GlossaError> {
-        writeln!(
-            self.output,
-            "{} «{}»",
-            "δοκιμή".bold().green(),
-            decl.name.as_str().italic()
-        )
-        .unwrap();
-
-        for stmt in &decl.body {
-            self.output.push_str("  ");
-            self.highlight_statement(stmt)?;
-            self.output.push('\n');
-        }
-
-        write!(self.output, "{}", "τέλος".bold()).unwrap();
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_highlight_simple_sentence() {
-        let source = "ὁ ἄνθρωπος τὸν λόγον λέγει.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-
-        // Check for ANSI codes
-        // Verify that words are present and some color codes are applied
-        assert!(output.contains("ἄνθρωπος"));
-        assert!(output.contains("\x1b[")); // Contains escape sequence
-    }
-
-    #[test]
-    fn test_highlight_string_literal() {
-        let source = "«χαῖρε» λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        // Italic (3) for string
-        assert!(output.contains("\x1b[3mχαῖρε\x1b[0m"));
-    }
-
-    #[test]
-    fn test_highlight_number_literal() {
-        let source = "42 λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        // Italic (3) for number
-        assert!(output.contains("\x1b[3m42\x1b[0m"));
-    }
-
-    #[test]
-    fn test_highlight_boolean_literal() {
-        let source = "ἀληθές λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        // Italic (3) for boolean
-        assert!(output.contains("\x1b[3mἀληθές\x1b[0m"));
-    }
-
-    #[test]
-    fn test_highlight_array_literal() {
-        let source = "[1, 2, 3] λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains('['));
-        assert!(output.contains(']'));
-        assert!(output.contains("\x1b[3m1\x1b[0m"));
-    }
-
-    #[test]
-    fn test_highlight_index_access() {
-        let source = "πίναξ[0] λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains('['));
-        assert!(output.contains(']'));
-        assert!(output.contains("\x1b[3m0\x1b[0m"));
-    }
-
-    #[test]
-    fn test_highlight_property_access() {
-        let source = "χρήστου ὄνομα λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        // Genitive owner (Magenta 35)
-        // Note: checking for color codes is brittle if crossterm changes, but magenta is typically 35
-        // We just ensure it runs and contains words
-        assert!(output.contains("χρήστου"));
-        assert!(output.contains("ὄνομα"));
-    }
-
-    #[test]
-    fn test_highlight_function_call() {
-        let source = "λέγε «χαῖρε».";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        // Verb Green (32)
-        assert!(output.contains("λέγε"));
-        assert!(output.contains("χαῖρε"));
-    }
-
-    #[test]
-    fn test_highlight_binding() {
-        let source = "ξ πέντε ἔστω.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("ξ"));
-        assert!(output.contains("πέντε")); // Numeral -> Italic
-        assert!(output.contains("\x1b[1mἔστω\x1b[0m")); // Bold
-    }
-
-    #[test]
-    fn test_highlight_binop() {
-        let source = "1 καὶ 2 λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        // Phrase(1, καὶ, 2) -> highlight words
-        assert!(output.contains("καὶ"));
-    }
-
-    #[test]
-    fn test_highlight_unary_op() {
-        // Negation
-        let source = "οὐκ ἀληθές λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("οὐ"));
-        assert!(output.contains("ἀληθές"));
-
-        // Unwrap
-        let source = "τιμή! λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("!"));
-    }
-
-    #[test]
-    fn test_highlight_block() {
-        // Block as a statement must end with period
-        let source = "{ «χαῖρε» λέγε. }.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("{"));
-        assert!(output.contains("}"));
-        assert!(output.contains("λέγε"));
-    }
-
-    #[test]
-    fn test_highlight_phrase() {
-        // A nested phrase expression is typically (expr expr)
-        // But parser produces Phrase for sequences of words.
-        // Let's try a function call that parses as a phrase initially
-        let source = "πρόσθεσις 1 2 λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("πρόσθεσις"));
-        assert!(output.contains("1"));
-        assert!(output.contains("2"));
-    }
-
-    #[test]
-    fn test_highlight_participle() {
-        let source = "τὰ διπλασιαζόμενα.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        // Participle Cyan (36)
-        assert!(output.contains("διπλασιαζόμενα"));
-    }
-
-    #[test]
-    fn test_highlight_definitions() {
-        // Type definition
-        let source = "εἶδος Χρήστης ὁρίζειν { ὄνομα Ὄνομα . ἡλικία Ἀριθμός }.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("εἶδος"));
-        assert!(output.contains("Χρήστης"));
-
-        // Trait definition
-        let source = "χαρακτήρ Δεικτόν ὁρίζειν { δεῖ δεῖξαι }.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("χαρακτήρ"));
-        assert!(output.contains("Δεικτόν"));
-    }
-
-    #[test]
-    fn test_highlight_test_declaration() {
-        let source = "δοκιμή «δοκιμή 1». «χαῖρε» λέγε. τέλος.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("δοκιμή"));
-        assert!(output.contains("χαῖρε"));
-        assert!(output.contains("τέλος"));
-    }
-
-    #[test]
-    fn test_highlight_query() {
-        let source = "ξ?";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("?"));
-    }
-
-    #[test]
-    fn test_highlight_propagate() {
-        let source = "σφάλμα;";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains(";"));
-    }
-
-    #[test]
-    fn test_highlight_unary_neg() {
-        // -1 might parse as literal -1. Use variable to force unary op.
-        let source = "-ξ λέγε.";
-        let result = highlight(source);
-        // If this fails, parser might not support unary minus yet or syntax is different
-        // Assuming it works for now based on grammar check
-        if let Ok(output) = result {
-            assert!(output.contains("-"));
-            assert!(output.contains("ξ"));
-        }
-    }
-
-    #[test]
-    fn test_highlight_clause_separator() {
-        let source = "εἰ ἀληθές, «ναί» λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains(","));
-    }
-
-    #[test]
-    fn test_highlight_nested_phrase() {
-        // Phrase inside a phrase
-        let source = "(1 2) λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("1"));
-        assert!(output.contains("2"));
-    }
-
-    #[test]
-    fn test_highlight_trait_impl() {
-        let source = "εἶδος Τύπος τῷ Χαρακτήρ ἐμπίπτειν { δεῖξαι 1. }.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("Τύπος"));
-        assert!(output.contains("Χαρακτήρ"));
-        assert!(output.contains("τῷ"));
-    }
-
-    #[test]
-    fn test_highlight_multiple_statements() {
-        let source = "ξ 1 ἔστω. ξ λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("\n")); // Separated by newline
-    }
-
-    #[test]
-    fn test_highlight_article_context() {
-        let source = "τὸν λόγον λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        // "τὸν" (Accusative) should set context for "λόγον"
-        // Analysis for "λόγον" is ambiguous (Nom/Acc), but context should resolve to Acc (Red)
-        // We verify it runs and contains content. Visual verification is done by human.
-        assert!(output.contains("λόγον"));
-    }
-
-    #[test]
-    fn test_highlight_complex_nested() {
-        // Nested structure: { [1, 2] λέγε. }.
-        let source = "{ [1, 2] λέγε. }.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_manual_ast_nodes() {
-        // These nodes are not currently produced by the parser but exist in the AST.
-        // We test them manually to ensure the highlighter handles them (future-proofing).
-        let mut h = Highlighter::new();
-
-        // 1. BinOp
-        let binop = Expr::BinOp {
-            left: Box::new(Expr::NumberLiteral(1)),
-            op: BinOperator::Add,
-            right: Box::new(Expr::NumberLiteral(2)),
-        };
-        h.highlight_expr(&binop).unwrap();
-        assert!(h.output.contains("+"));
-        h.output.clear();
-
-        // 2. Call
-        let call = Expr::Call {
-            verb: Word::new("λέγε"),
-            arguments: vec![Expr::StringLiteral("test".to_string())],
-        };
-        h.highlight_expr(&call).unwrap();
-        assert!(h.output.contains("λέγε"));
-        assert!(h.output.contains("test"));
-        h.output.clear();
-
-        // 3. Binding
-        let binding = Expr::Binding {
-            name: Word::new("χ"),
-            value: Box::new(Expr::NumberLiteral(10)),
-        };
-        h.highlight_expr(&binding).unwrap();
-        assert!(h.output.contains("χ"));
-        assert!(h.output.contains("ἔστω"));
-        h.output.clear();
-
-        // 4. PropertyAccess
-        let prop = Expr::PropertyAccess {
-            owner: Box::new(Expr::Word(Word::new("χρήστου"))),
-            property: Box::new(Expr::Word(Word::new("ὄνομα"))),
-        };
-        h.highlight_expr(&prop).unwrap();
-        assert!(h.output.contains("χρήστου"));
-        assert!(h.output.contains("ὄνομα"));
-        h.output.clear();
-
-        // 5. UnaryOp (Not)
-        let not_op = Expr::UnaryOp {
-            op: UnaryOperator::Not,
-            operand: Box::new(Expr::BooleanLiteral(true)),
-        };
-        h.highlight_expr(&not_op).unwrap();
-        assert!(h.output.contains("οὐ"));
-        h.output.clear();
-
-        // 6. UnaryOp (Neg)
-        let neg_op = Expr::UnaryOp {
-            op: UnaryOperator::Neg,
-            operand: Box::new(Expr::NumberLiteral(5)),
-        };
-        h.highlight_expr(&neg_op).unwrap();
-        assert!(h.output.contains("-"));
-        h.output.clear();
-    }
-
-    #[test]
-    fn test_highlight_pos_variants() {
-        // Test PartOfSpeech variants explicitly
-        let mut h = Highlighter::new();
-
-        // Preposition (white bold)
-        let prep = Word::new("μετά");
-        h.highlight_word(&prep).unwrap();
-        // Since we can't easily check ANSI codes for specific colors without fragile tests,
-        // we assume the logic works if we exercise the code path.
-        // We can check it's not empty.
-        assert!(!h.output.is_empty());
-        h.output.clear();
-
-        // Conjunction (white bold)
-        let conj = Word::new("καί");
-        h.highlight_word(&conj).unwrap();
-        assert!(!h.output.is_empty());
-        h.output.clear();
-
-        // Numeral (italic)
-        let num = Word::new("πέντε");
-        h.highlight_word(&num).unwrap();
-        assert!(!h.output.is_empty());
-        h.output.clear();
-
-        // Unknown (white)
-        let unknown = Word::new("ἀγνωστον");
-        h.highlight_word(&unknown).unwrap();
-        assert!(!h.output.is_empty());
-        h.output.clear();
-    }
-
-    #[test]
-    fn test_highlight_definitions_formatting() {
-        // Ensure bold/colors are applied to definition keywords
-        // Type - needs correct syntax: εἶδος Name ὁρίζειν { fields }
-        let source = "εἶδος Τ ὁρίζειν { α Α }.";
-        let res = highlight(source).unwrap();
-        // Check for bold escape sequence on εἶδος
-        assert!(res.contains("εἶδος"));
-        assert!(res.contains("\x1b[1m"));
-
-        // Trait - needs correct syntax: χαρακτήρ Name ὁρίζειν { methods }
-        let source = "χαρακτήρ Χ ὁρίζειν { δεῖ φ }.";
-        let res = highlight(source).unwrap();
-        assert!(res.contains("χαρακτήρ"));
-
-        // Impl - needs correct syntax: εἶδος Name τῷ Trait ἐμπίπτειν { methods }
-        let source = "εἶδος Τ τῷ Χ ἐμπίπτειν { φ 1. }.";
-        let res = highlight(source).unwrap();
-        // Missing "ἐμπίπτειν" in current output logic? Let's check.
-        // It outputs: εἶδος Τ τῷ Χ { ... } in highlight_trait_impl
-        // So we just check for εἶδος and τῷ
-        assert!(res.contains("εἶδος"));
-        assert!(res.contains("τῷ"));
-    }
-
-    #[test]
-    fn test_highlight_all_binops() {
-        let mut h = Highlighter::new();
-        let ops = [
-            BinOperator::Add,
-            BinOperator::Sub,
-            BinOperator::Mul,
-            BinOperator::Div,
-            BinOperator::Mod,
-            BinOperator::Eq,
-            BinOperator::Ne,
-            BinOperator::Lt,
-            BinOperator::Le,
-            BinOperator::Gt,
-            BinOperator::Ge,
-            BinOperator::And,
-            BinOperator::Or,
-        ];
-
-        for op in ops {
-            let expr = Expr::BinOp {
-                left: Box::new(Expr::NumberLiteral(1)),
-                op,
-                right: Box::new(Expr::NumberLiteral(2)),
-            };
-            h.output.clear();
-            h.highlight_expr(&expr).unwrap();
-            // Verify output is not empty (logic ran)
-            assert!(!h.output.is_empty());
-        }
-    }
-
-    #[test]
-    fn test_highlight_vocative_and_adjective() {
-        // "ἄνθρωπε" is Vocative singular of ἄνθρωπος
-        let source = "ἄνθρωπε λέγε.";
-        let result = highlight(source);
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert!(output.contains("ἄνθρωπε"));
-        // Vocative is Blue Italic
-        // Crossterm Italic is \x1b[3m
-        assert!(output.contains("\x1b[3m"));
-
-        // "καλός" is Adjective (Nominative Masculine)
-        // Adjective is Cyan
-        // Note: Crossterm might output `\x1b[36m` OR `\x1b[38;5;6m` (ANSI 256) depending on term detection
-        // But typically standard colors are simple codes.
-        // However, if test fails, let's just check for *any* color code that isn't white (which is default usually).
-        // Or better, let's trust that `crossterm::style::Stylize::cyan()` works and just check for escape.
-        let source_adj = "καλός.";
-        let result_adj = highlight(source_adj);
-        assert!(result_adj.is_ok());
-        let output_adj = result_adj.unwrap();
-        assert!(output_adj.contains("καλός"));
-        // Check for *some* coloring (at least [3...m)
-        assert!(output_adj.contains("\x1b[3"));
-    }
-}
