+++ src/lib.rs
@@ -46,7 +46,6 @@
 pub mod ast;
 pub mod codegen;
 pub mod errors;
-pub mod experimental;
 pub mod grammar;
 pub mod morphology;
 pub mod parser;
